pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chat-cli"
version = "1.0.0"
description = "一个智能聊天客户端CLI工具"
authors = [{name = "Your Name", email = "your.email@example.com"}]
requires-python = ">=3.8"
dependencies = [
    "typer[all]>=0.9.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "requests>=2.31.0",
    "colorama>=0.4.6",
    "openai>=1.3.0",
    "sse-starlette>=1.6.5",
    "python-dotenv>=1.0.0",
    "pydantic>=2.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[project.scripts]
chat = "main:main_entry"

[tool.setuptools.py-modules]
main = "main"
